Support to ADR for test release

+++ chirpstack/src/uplink/helpers.rs
@@ -44,7 +44,7 @@ pub fn get_uplink_dr(
         }
         chirpstack_api::gw::modulation::Parameters::Xss(v) => {
             lrwn::region::DataRateModulation::Xss(lrwn::region::XssDataRate {
-                spreading_factor: v.spreading_factor as u8,
+                spreading_factor: v.spreading_factor as u16,
                 bandwidth: v.bandwidth,
                 coding_rate: v.code_rate().into(),
             })
